+++ packages/features/di/modules/NoSlotsNotification.ts
@@ -1,5 +1,5 @@
-import type { INoSlotsNotificationService } from "@calcom/trpc/server/routers/viewer/slots/handleNotificationWhenNoSlots";
-import { NoSlotsNotificationService } from "@calcom/trpc/server/routers/viewer/slots/handleNotificationWhenNoSlots";
+import type { INoSlotsNotificationService } from "@calcom/features/slots/handleNotificationWhenNoSlots";
+import { NoSlotsNotificationService } from "@calcom/features/slots/handleNotificationWhenNoSlots";
 
 import { createModule } from "../di";
 import { DI_TOKENS } from "../tokens";
