+++ packages/features/slots/handleNotificationWhenNoSlots.test.ts
@@ -1,5 +1,5 @@
-import i18nMock from "../../../../../../tests/libs/__mocks__/libServerI18n";
-import prismaMock from "../../../../../../tests/libs/__mocks__/prismaMock";
+import i18nMock from "../../../tests/libs/__mocks__/libServerI18n";
+import prismaMock from "../../../tests/libs/__mocks__/prismaMock";
 
 import { vi, describe, it, beforeAll, afterAll, expect, beforeEach, afterEach } from "vitest";
