+++ packages/platform/libraries/slots.ts
@@ -2,7 +2,7 @@ import { FilterHostsService } from "@calcom/features/bookings/lib/host-filtering
 import { QualifiedHostsService } from "@calcom/features/bookings/lib/host-filtering/findQualifiedHostsWithDelegationCredentials";
 import { BusyTimesService } from "@calcom/features/busyTimes/services/getBusyTimes";
 import { validateRoundRobinSlotAvailability } from "@calcom/features/ee/round-robin/utils/validateRoundRobinSlotAvailability";
-import { NoSlotsNotificationService } from "@calcom/trpc/server/routers/viewer/slots/handleNotificationWhenNoSlots";
+import { NoSlotsNotificationService } from "@calcom/features/slots/handleNotificationWhenNoSlots";
 import { AvailableSlotsService } from "@calcom/trpc/server/routers/viewer/slots/util";
 
 export { AvailableSlotsService };
